move Utils::replace* to Replace class
